requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
charset-normalizer==2.1.0
click==8.1.3
docker-pycreds==0.4.0
frozenlist==1.3.0
fsspec==2022.7.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.48.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-service==2.4.0
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.9.0
setproctitle==1.3.0
shortuuid==1.0.9
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.8.1+cu101
torchaudio==0.8.1
torchmetrics==0.9.3
torchvision==0.9.1+cu101
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.12.21
Werkzeug==2.2.1
yarl==1.7.2
zipp==3.8.1
